wip: tidying up
