--- jdi-dark-tests/src/main/java/com/epam/jdi/services/websockets/WSGenericClient.java
@@ -1,9 +1,0 @@
-package com.epam.jdi.services.websockets;
-
-import com.epam.jdi.http.WebSocketGenericClient;
-
-import javax.websocket.ClientEndpoint;
-
-@ClientEndpoint
-public class WSGenericClient<T> extends WebSocketGenericClient<T> {
-}
